Move form provider to wizard level, with default values
